+++ kardia/client.go
@@ -25,7 +25,7 @@ import (
 	"github.com/kardiachain/go-kardia/lib/common"
 	"go.uber.org/zap"
 
-	"github.com/kardiachain/explorer-backend/types"
+	"github.com/kardiachain/kardia-explorer-backend/types"
 )
 
 type ClientInterface interface {
