+++ server/server.go
@@ -21,12 +21,11 @@ package server
 import (
 	"go.uber.org/zap"
 
-	"github.com/kardiachain/explorer-backend/types"
-
-	"github.com/kardiachain/explorer-backend/kardia"
-	"github.com/kardiachain/explorer-backend/metrics"
-	"github.com/kardiachain/explorer-backend/server/cache"
-	"github.com/kardiachain/explorer-backend/server/db"
+	"github.com/kardiachain/kardia-explorer-backend/cache"
+	"github.com/kardiachain/kardia-explorer-backend/db"
+	"github.com/kardiachain/kardia-explorer-backend/kardia"
+	"github.com/kardiachain/kardia-explorer-backend/metrics"
+	"github.com/kardiachain/kardia-explorer-backend/types"
 )
 
 type Config struct {
